[IMP]code improvement

+++ addons/website_sale/models/product.py
@@ -232,14 +232,15 @@ class Product(models.Model):
     @api.multi
     def write(self, vals):
         res = super(Product, self).write(vals)
-        for product in self:
-            product_image = product.product_product_image_ids.filtered('is_main_image')
-            if product_image and vals.get('image_variant') and product_image.image != vals['image_variant']:
-                product.product_product_image_ids = [(1, product_image.id, {'image': vals['image_variant']})]
-            elif product_image and 'image_variant' in vals:
-                product_image.unlink()
-            elif vals.get('image_variant'):
-                product.product_product_image_ids = [(0, 0, {'image': vals['image_variant'], 'name': self.name, 'is_main_image': True})]
+        if not self.env.context.get('no_image_update'):
+            for product in self:
+                product_image = product.product_product_image_ids.filtered('is_main_image')
+                if product_image and vals.get('image_variant') and product_image.image != vals['image_variant']:
+                    product.product_product_image_ids = [(1, product_image.id, {'image': vals['image_variant']})]
+                elif product_image and 'image_variant' in vals:
+                    product_image.unlink()
+                elif vals.get('image_variant'):
+                    product.product_product_image_ids = [(0, 0, {'image': vals['image_variant'], 'name': self.name, 'is_main_image': True})]
         return res
 
 
@@ -274,10 +275,17 @@ class ProductImage(models.Model):
                     product_image.product_tmpl_id.image = product_image.image
                 elif product_image.product_tmpl_id:
                     product_image.product_tmpl_id.write({'image_medium': False, 'image_small': False})
+                elif vals.get('image') and product_image.product_product_id:
+                    product_image.product_product_id.with_context(no_image_update=True).write({'image_variant': product_image.image})
+                elif product_image.product_product_id:
+                    product_image.product_product_id.with_context(no_image_update=True).write({'image_variant': False})
         return res
 
     @api.multi
     def unlink(self):
         for product_image in self.filtered(lambda p: p.is_main_image):
-            product_image.product_tmpl_id.write({'image_medium': False, 'image_small': False})
+            if product_image.product_tmpl_id:
+                product_image.product_tmpl_id.write({'image_medium': False, 'image_small': False})
+            elif product_image.product_product_id:
+                product_image.product_product_id.with_context(no_image_update=True).write({'image_variant': False})
         return super(ProductImage, self).unlink()
